+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added support for query `PeerVersion`
 - Added support for query `NodeInfo`
 - Added support for query `StartBaker`
+- Added support for query `StopBaker`
 
 ## 3.0.0
 - Stronger typing for various places in the API.